Add SettingService (no more Reactive)
alt title: Give up on Reactive, do it the event way

Reactive sounds really great, and the way everything is done now feels dirty, with a lot of event coupling, but it is a solution that works.

An implementation of the SettingsService is now finally added (and it works with the ShowGroups setting).